Rename notification hook files
