+++ ui/src/pages/namespace/Settings/Secrets/index.tsx
@@ -21,7 +21,7 @@ import ItemRow from "../components/ItemRow";
 import PaginationProvider from "~/components/PaginationProvider";
 import { SecretSchemaType } from "~/api/secrets/schema";
 import { useDeleteSecret } from "~/api/secrets/mutate/delete";
-import { useNotifications } from "~/api/namespaceLinting/query/useNamespaceLinting";
+import { useNotifications } from "~/api/notifications/query/get";
 import { useSecrets } from "~/api/secrets/query/get";
 import { useTranslation } from "react-i18next";
